ams_digical/requirements.txt
asgiref==3.8.1
attrs==25.1.0
Django==4.2
django-cors-headers==4.6.0
django-colorfield
django-filter
django-polymorphic
django-phonenumber-field==8.0.0
django-registration-redux==2.13
djangorestframework==3.15.2
django-rest-polymorphic
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
phonenumbers==8.13.50
pytz
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
sqlparse==0.5.2
setuptools
typing_extensions==4.12.2
tzdata==2024.2
djangorestframework-simplejwt

